requirements.txt
streamlit>=1.28.0
yt-dlp>=2025.10.22
openai-whisper>=20231117
moviepy>=1.0.3
opencv-python-headless>=4.8.1
ffmpeg-python>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.66.0
pillow>=10.0.0
ftfy
regex
# CLIP installation optional - can skip for testing
git+https://github.com/openai/CLIP.git


